+++ cedar-policy-validator/src/human_schema/fmt.rs
@@ -62,6 +62,7 @@ impl<N: Display> Display for SchemaType<N> {
             SchemaType::Type(ty) => match ty {
                 SchemaTypeVariant::Boolean => write!(f, "__cedar::Bool"),
                 SchemaTypeVariant::Entity { name } => write!(f, "{name}"),
+                SchemaTypeVariant::EntityOrCommon { type_name } => write!(f, "{type_name}"),
                 SchemaTypeVariant::Extension { name } => write!(f, "__cedar::{name}"),
                 SchemaTypeVariant::Long => write!(f, "__cedar::Long"),
                 SchemaTypeVariant::Record {
@@ -183,6 +184,20 @@ impl NameCollisionsError {
 }
 
 /// Convert a [`SchemaFragment`] to a string containing human schema syntax
+///
+/// As of this writing, this existing code throws an error if any
+/// fully-qualified name in a non-empty namespace is a valid common type and
+/// also a valid entity type.
+//
+// Two notes:
+// 1) This check is more conservative than necessary. Schemas are allowed to
+// shadow an entity type with a common type declaration in the same namespace;
+// see RFCs 24 and 70. What the human syntax can't express is if, in that
+// situation, we then specifically refer to the shadowed entity type name.  But
+// it's harder to walk all type references than it is to walk all type
+// declarations, so the conservative code here is fine; we can always make it
+// less conservative in the future without breaking people.
+// 2) This code is also likely the cause of #1063; see that issue
 pub fn json_schema_to_custom_schema_str<N: Display>(
     json_schema: &SchemaFragment<N>,
 ) -> Result<String, ToHumanSchemaSyntaxError> {
@@ -192,8 +207,8 @@ pub fn json_schema_to_custom_schema_str<N: Display>(
             .entity_types
             .keys()
             .map(|ty_name| {
-                RawName::new(ty_name.clone())
-                    .qualify_with(name.as_ref())
+                RawName::new_from_unreserved(ty_name.clone())
+                    .qualify_with_name(name.as_ref())
                     .to_smolstr()
             })
             .collect();
@@ -201,8 +216,8 @@ pub fn json_schema_to_custom_schema_str<N: Display>(
             .common_types
             .keys()
             .map(|ty_name| {
-                RawName::new(ty_name.clone())
-                    .qualify_with(name.as_ref())
+                RawName::new_from_unreserved(ty_name.clone())
+                    .qualify_with_name(name.as_ref())
                     .to_smolstr()
             })
             .collect();