+++ cedar-policy-validator/src/human_schema/parser.rs
@@ -25,7 +25,7 @@ use thiserror::Error;
 use super::{
     ast::Schema,
     err::{self, ParseError, ParseErrors, SchemaWarning, ToJsonSchemaErrors},
-    to_json_schema::{custom_schema_to_json_schema, custom_type_to_json_type},
+    to_json_schema::custom_schema_to_json_schema,
 };
 use cedar_policy_core::extensions::Extensions;
 
@@ -93,15 +93,6 @@ pub enum HumanSyntaxParseErrors {
     #[error(transparent)]
     #[diagnostic(transparent)]
     JsonError(#[from] ToJsonSchemaErrors),
-}
-
-/// Parse a type, in human syntax, into a [`crate::SchemaType`]
-pub fn parse_type(
-    src: &str,
-    extensions: Extensions<'_>,
-) -> Result<crate::SchemaType<crate::RawName>, HumanSyntaxParseErrors> {
-    let ty = parse_collect_errors(&*TYPE_PARSER, grammar::TypeParser::parse, src)?;
-    Ok(custom_type_to_json_type(ty, extensions.clone())?)
 }
 
 /// Parse a schema fragment, in human syntax, into a [`crate::SchemaFragment`],